Fixe 1

+++ src/main/java/fr/HtSTeam/HtS/GameModes/UHC/Common/UHC.java
@@ -40,7 +40,7 @@ public class UHC implements GameMode {
 
 	private void teleport() {
 
-		int border = OptionRegister.borderOption.getValue();
+		int border = (int) OptionRegister.borderOption.getValue();
 
 		if (TeamBuilder.teamList.size() == 0) {
 			for (Player player : Bukkit.getOnlinePlayers()) {
